awssummit2024-snowflake-GenAI-demo/requirements.txt
aiohttp==3.9.2
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
awscli==1.29.85
backoff==2.2.1
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.85
botocore==1.31.85
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
# cryptography==42.0.4
dataclasses-json==0.6.1
docutils==0.16
exceptiongroup==1.1.3
fastapi==0.109.1
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.41
h11==0.14.0
httptools==0.6.0
huggingface-hub
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.1.11
langchain-experimental==0.0.24
langsmith
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
onnxruntime==1.16.0
oscrypto==1.3.0
overrides==7.4.0
packaging==23.2
pandas==2.1.1
Pillow==10.2.0
pip==23.2.1
platformdirs==3.8.1
posthog==3.0.2
protobuf==4.24.3
pulsar-client==3.3.0
pyarrow==14.0.1
pyasn1==0.5.0
pycparser==2.21
pycryptodomex==3.19.1
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.6.0
rpds-py==0.10.3
rsa==4.7.2
s3transfer==0.7.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==65.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowflake-connector-python==3.2.0
snowflake-sqlalchemy==1.5.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.49
starlette
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
# tokenizers==0.13.3
toml==0.10.2
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
transformers==4.36.0
typer==0.9.0
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
watchfiles==0.20.0
websockets==11.0.3
wheel==0.38.4
yarl==1.9.2
zipp==3.17.0
aiohttp==3.9.2
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
awscli==1.29.85
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
bleach==6.0.0
blinker==1.6.2
boto3==1.34.75
botocore==1.34.75
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
comm==0.1.2
cryptography==41.0.7
curl_cffi==0.6.2
dataclasses-json==0.6.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docutils==0.16
duckduckgo_search==5.2.2
exceptiongroup==1.1.3
executing==1.2.0
Faker==17.6.0
fastapi==0.109.1
fastjsonschema==2.16.3
filelock==3.12.4
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.41
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-search-results==2.4.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.0
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
langchain==0.1.11
langchain-community==0.0.29
langchain-core==0.1.33
langchain-experimental==0.0.24
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.31
llama-index==0.10.26
llama-index-agent-openai==0.2.1
llama-index-cli==0.1.11
llama-index-core==0.10.26
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.14
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.13
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.0
llamaindex-py-client==0.1.15
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numexpr==2.8.7
numpy==1.26.0
onnxruntime==1.16.0
openai==1.16.0
orjson==3.10.0
oscrypto==1.3.0
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandasql==0.7.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.8.1
posthog==3.0.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.24.3
psutil==5.9.4
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodomex==3.19.1
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyOpenSSL==23.2.0
pyparsing==3.1.2
pypdf==4.1.0
PyPika==0.48.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.0.1
qtconsole==5.4.1
QtPy==2.3.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.3
rsa==4.7.2
s3transfer==0.10.1
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowflake-connector-python==3.2.0
snowflake-sqlalchemy==1.5.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==1.4.49
stack-data==0.6.2
starlette==0.35.1
streamlit==1.30.0
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.36.0
typer==0.9.0
types-requests==2.31.0.20240402
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
watchfiles==0.20.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
widgetsnbextension==4.0.5
wrapt==1.16.0
yarl==1.9.2
zipp==3.17.0
